Abstracted catalog bookkeeping into its own function
