+++ GAS/gasPipeline.py
@@ -4,6 +4,7 @@ import subprocess
 import glob
 import warnings
 from astropy.time import Time
+from . import catalogs
 
 def fillAll(overwrite=False):
 
@@ -14,7 +15,7 @@ def fillAll(overwrite=False):
         warnings.warn("fillAll() must be run on GB machines with access to lustre")
         return
 
-    updateLogs()
+    catalogs.updateLogs(release=release)
     catalog = parseLog()
     uniqSess = set(catalog['Session'].data.data)
     for session in uniqSess:
@@ -40,7 +41,7 @@ def fillAll(overwrite=False):
             subprocess.call(permissions,shell=True)
 
 def reduceAll(overwrite=False, release = 'all'):
-    updateLogs()
+    catalogs.updateLogs(release=release)
     catalog = parseLog()
     uniqSrc = set(catalog['Region name'].data.data)
     cwd = os.getcwd()
@@ -89,7 +90,7 @@ def wrapper(logfile='ObservationLog.csv',region='NGC1333',
     StartDate = Time(startdate)
     EndDate = Time(enddate)
     if not os.access(logfile,os.R_OK):
-        updateLogs()
+        catalogs.updateLogs(release=release)
 
     t = parseLog(logfile=logfile)
     for observation in t:
@@ -118,52 +119,6 @@ def wrapper(logfile='ObservationLog.csv',region='NGC1333',
                            Region=region,
                            RawDataDir=observation['Special RawDir'],
                            Window=str(thisWindow),overwrite=overwrite)
-
-
-
-def parseLog(logfile='ObservationLog.csv'):
-    """
-    Ingests a CSV log file into an astropy table
-    """
-    try:
-        from astropy.table import Table
-    except:
-        warnings.warn('GAS Pipeline requires astropy.  Try Anaconda!')
-        return
-    t = Table.read(logfile)
-    return(t)
-
-def updateLogs(output='ObservationLog.csv',release=None):
-    if release is None:
-        command = "wget --no-check-certificate --output-document="+output+" 'https://docs.google.com/spreadsheet/ccc?key=1F6MnXjK1Y1VyM8zWW3R5VvLAFF2Hkc85SGBRBxQ24JY&output=csv'"
-        # The returns from subprocess are the error codes from the OS
-        # If 0 then it worked so we should return True
-        return not subprocess.call(command,shell=True)
-    if 'DR1' in release:
-        from astropy.utils.data import get_pkg_data_filename
-        filename = get_pkg_data_filename('data/ObservationLog_DR1.csv',
-                                         package='GAS')
-        command = 'cp '+filename+' ./ObservationLog.csv'
-        return not subprocess.call(command,shell=True)
-    else:
-        warnings.warn('Updating logs failed for non-existent data release.')
-        return False
-
-def updateCatalog(output='RegionCatalog.csv',release=None):
-    if release is None:
-        command = "wget --no-check-certificate --output-document="+output+" 'https://docs.google.com/spreadsheets/d/140SUALscsm4Lco2WU3jDaREtUnf4jA9ZEBrMg4VAdKw/export?gid=1599734490&format=csv'"
-        return not subprocess.call(command,shell=True)
-    if 'DR1' in release:
-        from astropy.utils.data import get_pkg_data_filename
-        filename = get_pkg_data_filename('data/RegionCatalog_DR1.csv',
-                                         package='GAS')
-        command = 'cp '+filename+' ./'+output
-        return not subprocess.call(command,shell=True)
-    else:
-        warnings.warn('Updating logs failed for non-existent data release.')
-        return False
-
-
 
 
 def doPipeline(SessionNumber=1,StartScan = 11, EndScan=58,
